Add SampleSet tests

+++ test/analysis/metrics.jl
@@ -1,12 +1,12 @@
 function test_metrics()
     @testset "-*- Metrics" verbose = true begin
         e = 1.0
-        s = QUBOTools.SampleSet{Int,Float64}(
-            QUBOTools.Sample{Int,Float64}[
-                QUBOTools.Sample{Int,Float64}([0, 0, 1], 1, 1.0),
-                QUBOTools.Sample{Int,Float64}([0, 1, 0], 2, 2.0),
-                QUBOTools.Sample{Int,Float64}([0, 1, 1], 3, 3.0),
-                QUBOTools.Sample{Int,Float64}([1, 0, 0], 4, 4.0),
+        s = QUBOTools.SampleSet{Float64,Int}(
+            QUBOTools.Sample{Float64,Int}[
+                QUBOTools.Sample{Float64,Int}([0, 0, 1], 1, 1.0),
+                QUBOTools.Sample{Float64,Int}([0, 1, 0], 2, 2.0),
+                QUBOTools.Sample{Float64,Int}([0, 1, 1], 3, 3.0),
+                QUBOTools.Sample{Float64,Int}([1, 0, 0], 4, 4.0),
             ],
             Dict{String,Any}(
                 "time" => Dict{String,Any}(
@@ -23,15 +23,15 @@ function test_metrics()
             @test QUBOTools.success_rate(s, e) ≈ 0.1 atol = 1e-8
             @test QUBOTools.tts(s, e) ≈ 43.708690653 atol = 1e-8
 
-            let s = QUBOTools.SampleSet{Int,Float64}()
+            let s = QUBOTools.SampleSet{Float64,Int}()
                 @test isnan(QUBOTools.total_time(s))
                 @test isnan(QUBOTools.effective_time(s))
                 @test isnan(QUBOTools.success_rate(s, 0.0))
                 @test isnan(QUBOTools.tts(s, 0.0))
             end
 
-            let s = QUBOTools.SampleSet{Int,Float64}(
-                QUBOTools.Sample{Int,Float64}[],
+            let s = QUBOTools.SampleSet{Float64,Int}(
+                QUBOTools.Sample{Float64,Int}[],
                 Dict{String,Any}("time" => Dict{String,Any}()),
             )
                 @test isnan(QUBOTools.total_time(s))
@@ -40,8 +40,8 @@ function test_metrics()
                 @test isnan(QUBOTools.tts(s, 0.0))
             end
 
-            let s = QUBOTools.SampleSet{Int,Float64}(
-                QUBOTools.Sample{Int,Float64}[],
+            let s = QUBOTools.SampleSet{Float64,Int}(
+                QUBOTools.Sample{Float64,Int}[],
                 Dict{String,Any}(
                     "time" => Dict{String,Any}(
                         "total" => 1.0,
